Cache YAML swagger files
When not in a development environment we should cache the results of
compiling the swagger docs.
