+++ src/Ilios/ApiBundle/Resources/config/services.yml
@@ -26,3 +26,7 @@ services:
     ilios_api.endpoint_response_namer:
         class: Ilios\ApiBundle\Service\EndpointResponseNamer
         arguments: ["@kernel"]
+
+    ilios_api.swagger_doc_builder:
+        class: Ilios\ApiBundle\Service\SwaggerDocBuilder
+        arguments: ["@kernel", "%forceProtocol%"]
